Theme generator refactored
