+++ src/docs/DocsThemer/fonts.ts
@@ -1,0 +1,17 @@
+// Font Family - Translated Values
+
+export const fonts: any = {
+	// Tailwind Sans-Serif
+	sans: `Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'`,
+
+	// Tailwind Serif
+	serif: `ui-serif, Georgia, Cambria, "Times New Roman", Times, serif`,
+
+	// Tailwind Mono
+	mono: `ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace`,
+
+	// Helvetica
+	helvetica: `Helvetica, system-ui`
+
+	// TODO: extend with more options!
+};
